+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.5 (unreleased)
 
 - Added support for Rover
+- Improved error message when not fit
 
 ## 0.1.4 (2020-02-11)